separated bench module into benchfn and benchzstd
it shall be possible to use benchfn
without any dependency on zstd.
